+++ src/main/java/com/example/planj/frame/JoinFrame.java
@@ -1,4 +1,5 @@
-package com.example.planj;
+package com.example.planj.frame;
+
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.MouseAdapter;
