+++ src/main/java/com/example/planj/user/UserController.java
@@ -1,4 +1,4 @@
-package com.example.planj;
+package com.example.planj.user;
 
 
 import org.springframework.beans.factory.annotation.Autowired;
