+++ src/main/java/com/example/planj/user/UserRepository.java
@@ -1,4 +1,4 @@
-package com.example.planj;
+package com.example.planj.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
