+++ src/main/java/com/example/planj/user/UserService.java
@@ -1,6 +1,5 @@
-package com.example.planj;
+package com.example.planj.user;
 
-import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
